refactor: migrate netlify tests to typescript
