+++ packages/integrations/netlify/package.json
@@ -33,10 +33,11 @@
     "build": "astro-scripts build \"src/**/*.ts\" && tsc",
     "build:ci": "astro-scripts build \"src/**/*.ts\"",
     "test": "pnpm run test-fn && pnpm run test-static && pnpm run test:dev",
-    "test-fn": "astro-scripts test \"test/functions/*.test.js\"",
-    "test:dev": "astro-scripts test \"test/development/*.test.js\"",
-    "test-static": "astro-scripts test \"test/static/*.test.js\"",
-    "test:hosted": "astro-scripts test \"test/hosted/*.test.js\""
+    "test-fn": "astro-scripts test \"test/functions/*.test.ts\"",
+    "test:dev": "astro-scripts test \"test/development/*.test.ts\"",
+    "test-static": "astro-scripts test \"test/static/*.test.ts\"",
+    "test:hosted": "astro-scripts test \"test/hosted/*.test.js\"",
+    "typecheck:tests": "tsc --build tsconfig.test.json"
   },
   "dependencies": {
     "@astrojs/internal-helpers": "workspace:*",
