+++ packages/integrations/netlify/test/functions/skew-protection.test.ts
@@ -1,0 +1,64 @@
+import * as assert from 'node:assert/strict';
+import { readFile } from 'node:fs/promises';
+import { before, describe, it } from 'node:test';
+import { type Fixture, loadFixture } from '../../../../astro/test/test-utils.js';
+
+describe('Skew Protection', { timeout: 120000 }, () => {
+	let fixture: Fixture;
+
+	before(async () => {
+		// Set DEPLOY_ID env var for the test
+		process.env.DEPLOY_ID = 'test-deploy-123';
+
+		fixture = await loadFixture({
+			root: new URL('./fixtures/skew-protection/', import.meta.url),
+		});
+		await fixture.build({});
+
+		// Clean up
+		delete process.env.DEPLOY_ID;
+	});
+
+	it('Server islands inline adapter headers', async () => {
+		// Render a page with server islands and check the HTML contains inline headers
+		const entryURL = new URL(
+			'./fixtures/skew-protection/.netlify/v1/functions/ssr/ssr.mjs',
+			import.meta.url,
+		);
+		const { default: handler } = await import(entryURL.href);
+		const resp = await handler(new Request('http://example.com/server-island'), {});
+		const html = await resp.text();
+
+		// Check that the HTML contains the inline headers in the server island script
+		// Should have something like: const headers = new Headers({"X-Netlify-Deploy-ID":"test-deploy-123"});
+		assert.ok(
+			html.includes('test-deploy-123'),
+			'Expected server island HTML to include deploy ID in inline script',
+		);
+	});
+
+	it('Manifest contains internalFetchHeaders', async () => {
+		// The manifest is embedded in the build output
+		// Check the manifest file which contains the serialized manifest
+		const manifestURL = new URL(
+			'./fixtures/skew-protection/.netlify/build/chunks/',
+			import.meta.url,
+		);
+
+		// Find the manifest file (it has a hash in the name)
+		const { readdir } = await import('node:fs/promises');
+		const files = await readdir(manifestURL);
+		let found = false;
+		for (const file of files) {
+			const contents = await readFile(new URL(file, manifestURL), 'utf-8');
+			if (contents.includes('"internalFetchHeaders":{"X-Netlify-Deploy-ID":"test-deploy-123"}')) {
+				found = true;
+				break;
+			}
+		}
+		assert.ok(
+			found,
+			'Manifest should include internalFetchHeaders field with the correct deploy ID value',
+		);
+	});
+});
