+++ packages/integrations/netlify/tsconfig.test.json
@@ -1,0 +1,14 @@
+{
+  "extends": "../../../tsconfig.base.json",
+  "include": ["test/**/*.ts"],
+  "exclude": ["test/fixtures/**", "test/*/fixtures/**", "test/hosted/hosted-astro-project/**"],
+  "compilerOptions": {
+    "noEmit": true,
+    "allowJs": true,
+    "noUnusedLocals": false,
+    "noUnusedParameters": false,
+    "rewriteRelativeImportExtensions": true,
+    "rootDir": "."
+  },
+  "references": [{ "path": "../../astro/tsconfig.test.json" }]
+}
